Consolidate loader logic
